api/requirements.txt
fastapi
uvicorn
tensorflow
pydantic
prometheus_client
requests
Pillow
python-multipart
pytest
numpy